+++ common/rmi/src/main/java/is/codion/common/rmi/server/AbstractServer.java
@@ -13,6 +13,7 @@ import is.codion.common.user.User;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import sun.misc.ObjectInputFilter;
 
 import java.rmi.NoSuchObjectException;
 import java.rmi.Remote;
@@ -35,6 +36,7 @@ import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Predicate;
 
+import static is.codion.common.NullOrEmpty.nullOrEmpty;
 import static is.codion.common.rmi.server.RemoteClient.remoteClient;
 import static java.util.Collections.unmodifiableCollection;
 import static java.util.Objects.requireNonNull;
@@ -227,8 +229,9 @@ public abstract class AbstractServer<T extends Remote, A extends ServerAdmin> ex
     sharedAuthenticators.forEach(AbstractServer::closeAuthenticator);
     authenticators.values().forEach(AbstractServer::closeAuthenticator);
     auxiliaryServers.forEach(AbstractServer::stopAuxiliaryServer);
-    if (SerializationWhitelist.serializationDryRun()) {
-      SerializationWhitelist.writeDryRunWhitelist();
+    ObjectInputFilter serialFilter = ObjectInputFilter.Config.getSerialFilter();
+    if (serialFilter instanceof SerializationWhitelist.DryRun) {
+      ((SerializationWhitelist.DryRun) serialFilter).writeToFile(configuration.serializationFilterWhitelist());
     }
     shutdownEvent.run();
   }
@@ -446,11 +449,19 @@ public abstract class AbstractServer<T extends Remote, A extends ServerAdmin> ex
   }
 
   private static void configureSerializationWhitelist(ServerConfiguration configuration) {
-    if (configuration.serializationFilterDryRun()) {
-      SerializationWhitelist.configureDryRun(configuration.serializationFilterWhitelist());
+    String whitelistFile = configuration.serializationFilterWhitelist();
+    if (nullOrEmpty(whitelistFile)) {
+      LOG.info("No serialization whitelist file specified");
     }
     else {
-      SerializationWhitelist.configure(configuration.serializationFilterWhitelist());
+      if (configuration.serializationFilterDryRun()) {
+        ObjectInputFilter.Config.setSerialFilter(SerializationWhitelist.whitelistDryRun());
+        LOG.info("Serialization filter dry-run enabled");
+      }
+      else {
+        ObjectInputFilter.Config.setSerialFilter(SerializationWhitelist.whitelistFilter(whitelistFile));
+        LOG.info("Serialization filter whitelist set: " + whitelistFile);
+      }
     }
   }