+++ common/rmi/src/test/java/is/codion/common/rmi/server/SerializationWhitelistTest.java
@@ -4,7 +4,7 @@
 package is.codion.common.rmi.server;
 
 import is.codion.common.Serializer;
-import is.codion.common.rmi.server.SerializationWhitelist.SerializationFilter;
+import is.codion.common.rmi.server.SerializationWhitelist.WhitelistFilter;
 
 import org.junit.jupiter.api.Test;
 import sun.misc.ObjectInputFilter;
@@ -23,16 +23,15 @@ public final class SerializationWhitelistTest {
 
   @Test
   void dryRun() throws IOException, ClassNotFoundException {
-    assertThrows(IllegalArgumentException.class, () -> SerializationWhitelist.configureDryRun("classpath:dryrun"));
+    assertThrows(IllegalArgumentException.class, () -> SerializationWhitelist.whitelistDryRun().writeToFile("classpath:dryrun"));
     File tempFile = File.createTempFile("serialization_dry_run_test", "txt");
-    assertThrows(IllegalArgumentException.class, () -> SerializationWhitelist.configureDryRun(tempFile.getAbsolutePath()));
-    tempFile.delete();
 
-    SerializationWhitelist.configureDryRun(tempFile.getAbsolutePath());
+    SerializationWhitelist.DryRun serialFilter = SerializationWhitelist.whitelistDryRun();
+    ObjectInputFilter.Config.setSerialFilter(serialFilter);
 
     Serializer.deserialize(Serializer.serialize(Integer.valueOf(42)));
     Serializer.deserialize(Serializer.serialize(Long.valueOf(42)));
-    SerializationWhitelist.writeDryRunWhitelist();
+    serialFilter.writeToFile(tempFile.getAbsolutePath());
 
     List<String> classNames = Files.readAllLines(tempFile.toPath(), StandardCharsets.UTF_8);
 
@@ -42,7 +41,7 @@ public final class SerializationWhitelistTest {
     assertEquals(Number.class.getName(), classNames.get(2));
 
     Serializer.deserialize(Serializer.serialize(Double.valueOf(42)));
-    SerializationWhitelist.writeDryRunWhitelist();
+    serialFilter.writeToFile(tempFile.getAbsolutePath());
 
     classNames = Files.readAllLines(tempFile.toPath(), StandardCharsets.UTF_8);
 
@@ -63,29 +62,29 @@ public final class SerializationWhitelistTest {
             "is.codion.common.state.State",
             "is.codion.common.state.StateObserver"
     );
-    SerializationFilter filter = new SerializationFilter(whitelistItems);
-    assertEquals(filter.checkInput("is.codion.common.value.Value"), sun.misc.ObjectInputFilter.Status.ALLOWED);
-    assertEquals(filter.checkInput("is.codion.common.state.State"), sun.misc.ObjectInputFilter.Status.ALLOWED);
+    WhitelistFilter filter = SerializationWhitelist.whitelistFilter(whitelistItems);
+    assertEquals(filter.checkInput("is.codion.common.value.Value"), ObjectInputFilter.Status.ALLOWED);
+    assertEquals(filter.checkInput("is.codion.common.state.State"), ObjectInputFilter.Status.ALLOWED);
     assertEquals(filter.checkInput("is.codion.common.state.States"), ObjectInputFilter.Status.REJECTED);
-    assertEquals(filter.checkInput("is.codion.common.state.StateObserver"), sun.misc.ObjectInputFilter.Status.ALLOWED);
+    assertEquals(filter.checkInput("is.codion.common.state.StateObserver"), ObjectInputFilter.Status.ALLOWED);
     assertEquals(filter.checkInput("is.codion.common.event.Event"), ObjectInputFilter.Status.REJECTED);
     assertEquals(filter.checkInput("is.codion.common.i18n.Messages"), ObjectInputFilter.Status.REJECTED);
   }
 
   @Test
   void file() {
-    assertThrows(RuntimeException.class, () -> new SerializationFilter("src/test/resources/whitelist_test_non_existing.txt"));
-    testFilter(new SerializationFilter("src/test/resources/whitelist_test.txt"));
+    assertThrows(RuntimeException.class, () -> SerializationWhitelist.whitelistFilter("src/test/resources/whitelist_test_non_existing.txt"));
+    testFilter(SerializationWhitelist.whitelistFilter("src/test/resources/whitelist_test.txt"));
   }
 
   @Test
   void classpath() {
-    assertThrows(IllegalArgumentException.class, () -> new SerializationFilter("classpath:src/test/resources/whitelist_test.txt"));
-    testFilter(new SerializationFilter("classpath:whitelist_test.txt"));
-    testFilter(new SerializationFilter("classpath:/whitelist_test.txt"));
+    assertThrows(IllegalArgumentException.class, () -> SerializationWhitelist.whitelistFilter("classpath:src/test/resources/whitelist_test.txt"));
+    testFilter(SerializationWhitelist.whitelistFilter("classpath:whitelist_test.txt"));
+    testFilter(SerializationWhitelist.whitelistFilter("classpath:/whitelist_test.txt"));
   }
 
-  private static void testFilter(SerializationFilter filter) {
+  private static void testFilter(WhitelistFilter filter) {
     assertEquals(filter.checkInput("is.codion.common.value.Value"), ObjectInputFilter.Status.ALLOWED);
     assertEquals(filter.checkInput("is.codion.common.state.State"), ObjectInputFilter.Status.ALLOWED);
     assertEquals(filter.checkInput("is.codion.common.state.States"), ObjectInputFilter.Status.ALLOWED);