+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - AbstractDatabase, login timeout now set in a static initializer block, instead of during connection creation.
 ### is.codion.common.rmi
 - SerializationWhitelist.writeToFile() improved a bit, javadocs fixed.
+- SerializationWhitelist no longer public, refactored.
 ### is.codion.swing.common.ui
 - KeyboardShortcuts.copy() added.
 - CalendarPanel.Builder added, replaces factory methods